feat: remove old errors
