+++ packages/errors/src/helpers/helpers.ts
@@ -31,6 +31,27 @@ const stringifyData = (data: unknown): string => {
 };
 
 /**
+ * Asserts that the given error is an instance of the Error class.
+ * If the error is an instance of Error, it is returned.
+ * If the error is not an instance of Error, a new Error object is created with a descriptive message.
+ *
+ * @param {unknown} error - The error to be asserted.
+ * @return {Error} - The error if it is an instance of Error, or a new Error object if it is not.
+ *
+ * @remarks
+ * **IMPORTANT: no sensitive data should be passed as any parameter.**
+ */
+function assertInnerError(error: unknown): Error {
+    if (error instanceof Error) {
+        return error;
+    }
+
+    return new Error(
+        `Inner error is not an instance of Error. Object:\n\t${stringifyData(error)}`
+    );
+}
+
+/**
  * Function to build the error message.
  * Here, we can customize the error message format.
  *
@@ -54,4 +75,4 @@ function createErrorMessage<TErrorDataType>(
     );
 }
 
-export { createErrorMessage };
+export { stringifyData, assertInnerError, createErrorMessage };
