--- packages/errors/src/model/core/index.ts
@@ -1,11 +1,0 @@
-export * from './abi';
-export * from './address';
-export * from './bloom';
-export * from './certificate';
-export * from './contract';
-export * from './data';
-export * from './hdnode';
-export * from './keystore';
-export * from './rlp';
-export * from './secp256k1';
-export * from './transaction';
