--- packages/errors/src/model/eip1193/index.ts
@@ -1,1 +1,0 @@
-export * from './eip1193';
