--- packages/errors/src/model/generic/function.ts
@@ -1,19 +1,0 @@
-import { type DefaultErrorData } from '../../types';
-import { ErrorBase } from '../base';
-
-/**
- * Not implemented function
- */
-class NotImplementedError extends ErrorBase<
-    FUNCTION.NOT_IMPLEMENTED,
-    DefaultErrorData
-> {}
-
-/**
- * Errors enum.
- */
-enum FUNCTION {
-    NOT_IMPLEMENTED = 'NOT_IMPLEMENTED'
-}
-
-export { NotImplementedError, FUNCTION };
