--- packages/errors/src/model/generic/index.ts
@@ -1,1 +1,0 @@
-export * from './function';
