--- packages/errors/src/model/index.ts
@@ -1,6 +1,0 @@
-export * from './base';
-export * from './core';
-export * from './network';
-export * from './generic';
-export * from './eip1193';
-export * from './provider-rpc';
