--- packages/errors/src/utils/assert/index.ts
@@ -1,1 +1,0 @@
-export * from './assert';
