--- packages/errors/src/utils/error-builder/index.ts
@@ -1,1 +1,0 @@
-export * from './error-builder';
