--- packages/errors/src/utils/error-message-builder/index.ts
@@ -1,1 +1,0 @@
-export * from './error-message-builder';
