--- packages/errors/src/utils/index.ts
@@ -1,3 +1,0 @@
-export * from './error-builder';
-export * from './error-message-builder';
-export * from './assert';
